backend/requirements.txt
fastapi
uvicorn
elasticsearch
pydantic
python-dotenv
google-generativeai
requests
